Rename common.d.ts to jquery_augmentation.d.ts
