+++ js/data/odata/context.d.ts
@@ -1,6 +1,6 @@
 import {
     JQueryPromise
-} from '../../common';
+} from '../../jquery_augmentation';
 
 export interface ODataContextOptions {
     /**
