+++ js/jquery_augmentation.d.ts
