+++ js/ui/context_menu.d.ts
@@ -5,7 +5,7 @@ import {
 import {
     JQueryEventObject,
     JQueryPromise
-} from '../common';
+} from '../jquery_augmentation';
 
 import {
     dxElement
