+++ js/ui/defer_rendering.d.ts
@@ -4,7 +4,7 @@ import {
 
 import {
     JQueryPromise
-} from '../common';
+} from '../jquery_augmentation';
 
 import {
     dxElement
