+++ js/ui/drop_down_editor/ui.drop_down_list.d.ts
@@ -1,6 +1,6 @@
 import {
     JQueryEventObject
-} from '../../common';
+} from '../../jquery_augmentation';
 
 import {
     dxElement
