+++ js/ui/pivot_grid_field_chooser.d.ts
@@ -1,6 +1,6 @@
 import {
     JQueryEventObject
-} from '../common';
+} from '../jquery_augmentation';
 
 import {
     dxElement
