+++ js/ui/popover.d.ts
@@ -8,7 +8,7 @@ import {
 
 import {
     JQueryPromise
-} from '../common';
+} from '../jquery_augmentation';
 
 import {
     event
