+++ js/ui/resizable.d.ts
@@ -1,6 +1,6 @@
 import {
     JQueryEventObject
-} from '../common';
+} from '../jquery_augmentation';
 
 import DOMComponent, {
     DOMComponentOptions
