+++ js/ui/select_box.d.ts
@@ -1,6 +1,6 @@
 import {
     JQueryPromise
-} from '../common';
+} from '../jquery_augmentation';
 
 import {
     dxElement
