+++ js/ui/tree_view.d.ts
@@ -1,7 +1,7 @@
 import {
     JQueryEventObject,
     JQueryPromise
-} from '../common';
+} from '../jquery_augmentation';
 
 import {
     dxElement
