+++ js/viz/chart_components/base_chart.d.ts
@@ -1,6 +1,6 @@
 import {
     JQueryEventObject
-} from '../../common';
+} from '../../jquery_augmentation';
 
 import {
     dxElement
